Bring back <GenerateInvoice />.
